Moves the possible mission status object to an exportable util
