+++ gui/src/utils/MissionStatuses.js
@@ -1,0 +1,8 @@
+const StatusEnum = Object.freeze({
+  STARTMISSION: "Waiting to Start",
+  UNVERIFIED: "Not Verified",
+  VERIFIED: "Verified",
+  ABORTED: "Aborted",
+});
+
+export default StatusEnum;
